requirements.txt
flask
flask-mysqldb
python-dotenv
werkzeug
email_validator
